Reorganise file, fix storage
