+++ docs/Duke/Deadline.html
@@ -123,10 +123,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="Task.html" title="class in duke">duke.Task</a></li>
+<li><a href="Task.html" title="class in duke">duke.common.task.Task</a></li>
 <li>
 <ul class="inheritance">
-<li>duke.Deadline</li>
+<li>duke.common.task.Deadline</li>
 </ul>
 </li>
 </ul>
@@ -144,7 +144,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 extends <a href="Task.html" title="class in duke">Task</a></pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../serialized-form.html#duke.Deadline">Serialized Form</a></dd>
+<dd><a href="../serialized-form.html#duke.common.task.Deadline">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
@@ -195,7 +195,7 @@ extends <a href="Task.html" title="class in duke">Task</a></pre>
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.duke.Task">
+<li class="blockList"><a id="methods.inherited.from.class.duke.common.task.Task">
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;duke.<a href="Task.html" title="class in duke">Task</a></h3>
