+++ docs/Duke/Duke.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Duke</li>
+<li>duke.common.Duke</li>
 </ul>
 </li>
 </ul>
