+++ docs/Duke/Parser.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Parser</li>
+<li>duke.common.Parser</li>
 </ul>
 </li>
 </ul>
@@ -176,7 +176,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>static boolean</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#parse(java.lang.String,duke.TaskList)">parse</a></span>&#8203;(java.lang.String&nbsp;input,
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#parse(java.lang.String,duke.common.TaskList)">parse</a></span>&#8203;(java.lang.String&nbsp;input,
      <a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;taskList)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
@@ -223,7 +223,7 @@ extends java.lang.Object</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="parse(java.lang.String,duke.TaskList)">
+<a id="parse(java.lang.String,duke.common.TaskList)">
 <!--   -->
 </a>
 <ul class="blockListLast">
