+++ docs/Duke/Storage.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Storage</li>
+<li>duke.common.Storage</li>
 </ul>
 </li>
 </ul>
@@ -181,7 +181,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#store(duke.TaskList)">store</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;taskList)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#store(duke.common.TaskList)">store</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;taskList)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 </table>
@@ -236,7 +236,7 @@ extends java.lang.Object</pre>
 <pre class="methodSignature">public&nbsp;<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;initialise()</pre>
 </li>
 </ul>
-<a id="store(duke.TaskList)">
+<a id="store(duke.common.TaskList)">
 <!--   -->
 </a>
 <ul class="blockListLast">
