+++ docs/Duke/Task.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Task</li>
+<li>duke.common.task.Task</li>
 </ul>
 </li>
 </ul>
@@ -144,7 +144,7 @@ extends java.lang.Object
 implements java.io.Serializable</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../serialized-form.html#duke.Task">Serialized Form</a></dd>
+<dd><a href="../serialized-form.html#duke.common.task.Task">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
