+++ docs/Duke/TaskList.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.TaskList</li>
+<li>duke.common.TaskList</li>
 </ul>
 </li>
 </ul>
@@ -140,7 +140,7 @@ extends java.lang.Object
 implements java.io.Serializable</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../serialized-form.html#duke.TaskList">Serialized Form</a></dd>
+<dd><a href="../serialized-form.html#duke.common.TaskList">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
@@ -189,7 +189,7 @@ implements java.io.Serializable</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#addCustom(duke.Task)">addCustom</a></span>&#8203;(<a href="Task.html" title="class in duke">Task</a>&nbsp;task)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#addCustom(duke.common.task.Task)">addCustom</a></span>&#8203;(<a href="Task.html" title="class in duke">Task</a>&nbsp;task)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr id="i1" class="rowColor">
@@ -264,7 +264,7 @@ implements java.io.Serializable</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="addCustom(duke.Task)">
+<a id="addCustom(duke.common.task.Task)">
 <!--   -->
 </a>
 <ul class="blockList">
