+++ docs/Duke/Ui.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Ui</li>
+<li>duke.common.Ui</li>
 </ul>
 </li>
 </ul>
