+++ docs/index-files/index-11.html
@@ -105,7 +105,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../StorageTest.html#%3Cinit%3E()">StorageTest()</a></span> - Constructor for class <a href="../StorageTest.html" title="class in &lt;Unnamed&gt;">StorageTest</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../duke/Storage.html#store(duke.TaskList)">store(TaskList)</a></span> - Method in class duke.<a href="../duke/Storage.html" title="class in duke">Storage</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Storage.html#store(duke.common.TaskList)">store(TaskList)</a></span> - Method in class duke.<a href="../duke/Storage.html" title="class in duke">Storage</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">P</a>&nbsp;<a href="index-10.html">R</a>&nbsp;<a href="index-11.html">S</a>&nbsp;<a href="index-12.html">T</a>&nbsp;<a href="index-13.html">U</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
