+++ docs/index-files/index-9.html
@@ -97,7 +97,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">P</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../duke/Parser.html#parse(java.lang.String,duke.TaskList)">parse(String, TaskList)</a></span> - Static method in class duke.<a href="../duke/Parser.html" title="class in duke">Parser</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Parser.html#parse(java.lang.String,duke.common.TaskList)">parse(String, TaskList)</a></span> - Static method in class duke.<a href="../duke/Parser.html" title="class in duke">Parser</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../duke/Parser.html" title="class in duke"><span class="typeNameLink">Parser</span></a> - Class in <a href="../duke/package-summary.html">duke</a></dt>
 <dd>&nbsp;</dd>
