+++ docs/serialized-form.html
@@ -101,10 +101,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <section role="region">
 <h2 title="Package">Package&nbsp;duke</h2>
 <ul class="blockList">
-<li class="blockList"><a id="duke.Deadline">
+<li class="blockList"><a id="duke.common.task.Deadline">
 <!--   -->
 </a>
-<h3>Class <a href="duke/Deadline.html" title="class in duke">duke.Deadline</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
+<h3>Class <a href="duke/Deadline.html" title="class in duke">duke.common.task.Deadline</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
@@ -122,10 +122,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h3>Class <a href="duke/duke.DukeException.html" title="class in duke">duke.duke.DukeException</a> extends java.lang.Exception implements Serializable</h3>
 </li>
-<li class="blockList"><a id="duke.Event">
+<li class="blockList"><a id="duke.common.task.Event">
 <!--   -->
 </a>
-<h3>Class <a href="duke/Event.html" title="class in duke">duke.Event</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
+<h3>Class <a href="duke/Event.html" title="class in duke">duke.common.task.Event</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
@@ -138,10 +138,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="blockList"><a id="duke.Task">
+<li class="blockList"><a id="duke.common.task.Task">
 <!--   -->
 </a>
-<h3>Class <a href="duke/Task.html" title="class in duke">duke.Task</a> extends java.lang.Object implements Serializable</h3>
+<h3>Class <a href="duke/Task.html" title="class in duke">duke.common.task.Task</a> extends java.lang.Object implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
@@ -158,10 +158,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="blockList"><a id="duke.TaskList">
+<li class="blockList"><a id="duke.common.TaskList">
 <!--   -->
 </a>
-<h3>Class <a href="duke/TaskList.html" title="class in duke">duke.TaskList</a> extends java.lang.Object implements Serializable</h3>
+<h3>Class <a href="duke/TaskList.html" title="class in duke">duke.common.TaskList</a> extends java.lang.Object implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
@@ -178,10 +178,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="blockList"><a id="duke.Todo">
+<li class="blockList"><a id="duke.common.task.Todo">
 <!--   -->
 </a>
-<h3>Class <a href="duke/Todo.html" title="class in duke">duke.Todo</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
+<h3>Class <a href="duke/Todo.html" title="class in duke">duke.common.task.Todo</a> extends <a href="duke/Task.html" title="class in duke">Task</a> implements Serializable</h3>
 </li>
 </ul>
 </section>
