+++ src/main/java/duke/Main.java
@@ -2,6 +2,8 @@ package duke;
 
 import java.io.IOException;
 
+import duke.common.Duke;
+import duke.controller.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
