+++ src/main/java/duke/common/Parser.java
@@ -1,8 +1,12 @@
-package duke;
+package duke.common;
 
 import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import duke.common.task.Deadline;
+import duke.common.task.Event;
+import duke.common.task.Task;
+import duke.common.task.Todo;
 
 /**
  * Parses inputs provided by ui, and directs ui to print corresponding output.
