+++ src/main/java/duke/common/task/Deadline.java
@@ -1,8 +1,9 @@
-package duke;
+package duke.common.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import duke.common.task.Task;
 
 public class Deadline extends Task {
     private LocalDate deadline;
