+++ src/main/java/duke/common/task/Todo.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.common.task;
+
+import duke.common.task.Task;
 
 public class Todo extends Task {
     /**
